requirements.txt
Appium-Python-Client==4.5.0
PyYAML==6.0.1
pytest==7.4.4
pytest-xdist==3.5.0
pytest-rerunfailures==12.0
allure-pytest==2.13.2
